requirements.txt
pyTelegramBotAPI
google-generativeai
requests
beautifulsoup4
python-dotenv
